+++ src/Exceptionist.php
@@ -261,6 +261,19 @@ class Exceptionist
             }
         }
 
-        throw $exception instanceof Throwable ? $exception : new $exception($message);
+        if ($exception instanceof \Throwable) {
+            throw $exception;
+        }
+
+        //Tries to set file and line that throwned the exception
+        if ($exception == ErrorException::class || is_subclass_of($exception, ErrorException::class)) {
+            foreach (debug_backtrace() as $backtrace) {
+                if ($backtrace['file'] != __FILE__) {
+                    throw new $exception($message, 0, E_ERROR, $backtrace['file'], $backtrace['line']);
+                }
+            }
+        }
+
+        throw new $exception($message);
     }
 }