+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # 1.x branch
 ## 1.5 branch
 ### 1.5.0
-* all exception classes now extend the `ErrorException`;
+* the `Exceptionist` can now set file and line that throwned the exception. All
+    exception classes provided by `php-tools` now extend the `ErrorException`;
 * all filesystem global function have been removed, use `Filesystem` class instead;
 * `TestTrait::assertFilePerms()` has been removed. Use instead
     `assertFileIsReadable()`/`assertFileIsWritable()`/